Added 'verifyCommand' function to commands module, to check that only actual usable commands can have permissions set. Added 'isQuoted' check to utils module, to check if a given string is surrounded by either single or double quotes.
Completed first draft of the 'perms' module. The module introduces perm (permission) objects that can be applied to any of the other existing objects (users, places, aliases, and groups) as well as to commands.
The perms module includes functions for getting, adding, removing and updating perm records in the database.
The module initializes a 'perms' table in the database.
It also includes commands for adding, removing, editing and showing permission records through the command interface.